src/requirements.txt
# Web Scraping (Updated)
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Vector Databases (Alternatives)
chromadb==0.4.22
pinecone-client==3.0.0
pgvector==0.2.4

# Embeddings & ML
sentence-transformers==2.2.2
torch==2.1.0
numpy==1.24.3
scikit-learn==1.3.2

# LLM Integration (Gemini)
google-generativeai==0.3.2
langchain==0.1.0
langchain-community==0.0.13
langchain-google-genai==0.0.6

# Backend API (Phase 8)
fastapi==0.104.1
uvicorn[standard]==0.24.0
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Streamlit Deployment
streamlit==1.31.0

# Data Processing
pandas==2.1.3
pydantic==2.5.0
python-dateutil==2.8.2

# Environment Variables
python-dotenv==1.0.0

# Utilities
tqdm==4.66.1
retrying==1.3.4

# Testing
pytest==7.4.3
pytest-cov==4.1.0
coverage==7.4.0

# CLI
click==8.1.7
prompt-toolkit==3.0.41
